Refactor json_test_runner to modularize.
