+++ tests/column_map_expectations/expect_column_values_to_be_in_set.json
@@ -53,7 +53,7 @@
     {
       "notes": "Missing required parameters",
       "in": {},
-      "raises_exception": "TypeError"
+      "exception": "TypeError"
     }
   ]
 }
