+++ tests/util.py
@@ -4,6 +4,7 @@ import pandas as pd
 import numpy as np
 import os
 import json
+import pytest
 
 from sqlalchemy import create_engine
 
@@ -129,7 +130,52 @@ def discover_json_test_configurations(test_config_dir):
     return test_configurations, test_configuration_ids
 
 
+def evaluate_json_test(test_configuration, dataset_type):
+    # Get the test configuration
+    # with open(test_config_dir + test_configuration_file) as configuration_file:
+    #     test_configuration = json.load(configuration_file)
 
-def get_tests(test_name):
-    with open("./tests/test_sets/" + test_name + ".json") as f:
-        yield json.load(f)
+    # Build the required dataset
+    dataset = get_dataset(dataset_type, test_configuration['dataset'])
+    dataset.set_default_expectation_argument('result_format', 'COMPLETE')
+
+    # Get the expectation to run
+    expectation_name = test_configuration['expectation']
+
+    # Run the expectation and get results
+    # for test in test_configuration['tests']:
+    test = test_configuration['test']
+
+    if 'exception' in test:
+        with pytest.raises(Exception) as exception_info:
+            if isinstance(test['in'], list):
+                getattr(dataset, expectation_name)(*test['in'])
+            else:
+                result = getattr(dataset, expectation_name)(**test['in'])
+        assert test['exception'] in str(exception_info)
+
+    else:
+        # Support tests with positional arguments
+        if isinstance(test['in'], list):
+            result = getattr(dataset, expectation_name)(*test['in'])
+        # As well as keyword arguments
+        else:
+            result = getattr(dataset, expectation_name)(**test['in'])
+
+        # Check results
+        if 'out' in test:
+            assert result['success'] == test['out']['success']
+
+            # Handle dataset-specific implementation differences here
+            if 'unexpected_index_list' in test['out']:
+                if dataset_type == 'SqlAlchemyDataSet':
+                    pass
+                else:
+                    assert result['result_obj']['unexpected_index_list'] == test['out']['unexpected_index_list']
+
+            if 'unexpected_list' in test['out']:
+                assert result['result_obj']['unexpected_list'] == test['out']['unexpected_list']
+
+        if 'error' in test:
+            assert result['exception_info']['raised_exception'] == True
+            assert test['error']['traceback_substring'] in result['excpetion_info']['exception_traceback']
